--- client/src/components/auth/AuthLinks.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { useAuthActions } from "./AuthActions";
-
-export function SignoutLink() {
-  const { signOut } = useAuthActions();
-  return (
-    <button
-      type="button"
-      className="w-full text-left cursor-pointer"
-      onClick={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-        signOut();
-      }}
-    >
-      Sign Out
-    </button>
-  );
-}
